Make file structure better
