+++ next.config.mjs
@@ -1,6 +1,11 @@
 import withFlowbiteReact from "flowbite-react/plugin/nextjs";
 
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  images: {
+    domains: ["cdn.dummyjson.com"],
+  },
+};
 
-export default withFlowbiteReact(nextConfig);
+// export default withFlowbiteReact(nextConfig);
+export default nextConfig;
